Fix validate:benchmark:statistics and export:benchmark:statistics
