+++ src/Command/Validate/Benchmark/ValidateBenchmarkPhpInfoCommand.php
@@ -5,22 +5,15 @@ declare(strict_types=1);
 namespace App\Command\Validate\Benchmark;
 
 use App\{
-    Benchmark\Benchmark,
     Benchmark\BenchmarkUrlService,
     BenchmarkConfiguration\BenchmarkConfiguration,
-    BenchmarkConfiguration\BenchmarkConfigurationService,
-    Command\AbstractCommand,
     Command\Behavior\CallUrlTrait,
-    Command\Behavior\ValidateCircleCiOption,
-    Command\Benchmark\BenchmarkInitCommand,
-    Command\Validate\Configuration\ValidateConfigurationCommand,
     PhpVersion\PhpVersion
 };
 
-final class ValidateBenchmarkPhpInfoCommand extends AbstractCommand
+final class ValidateBenchmarkPhpInfoCommand extends AbstractValidateBenchmarkUrlCommand
 {
     use CallUrlTrait;
-    use ValidateCircleCiOption;
 
     /** @var string */
     protected static $defaultName = 'validate:benchmark:php-info';
@@ -29,62 +22,23 @@ final class ValidateBenchmarkPhpInfoCommand extends AbstractCommand
     {
         parent::configure();
 
-        $this
-            ->setDescription('Validate benchmark phpinfo()')
-            ->addOption('no-validate-configuration')
-            ->addValidateCircleCiOption($this->getDefinition());
+        $this->setDescription('Validate benchmark response');
     }
 
-    protected function doExecute(): int
+    protected function getUrl(): string
     {
-        if ($this->getInput()->getOption('no-validate-configuration') === false) {
-            $this->runCommand(
-                ValidateConfigurationCommand::getDefaultName(),
-                $this->appendValidateCircleCiOption($this->getInput())
-            );
-        }
-
-        foreach (Benchmark::getCompatiblesPhpVersions() as $phpVersion) {
-            $this->validateForPhpVersion($phpVersion);
-        }
-
-        return 0;
+        return BenchmarkUrlService::getPhpinfoUrl();
     }
 
-    protected function validateForPhpVersion(PhpVersion $phpVersion): self
-    {
-        foreach (BenchmarkConfigurationService::getAvailable($phpVersion) as $benchmarkConfiguration) {
-            $this->outputTitle(
-                'Validation of '
-                    . BenchmarkUrlService::getPhpinfoUrl()
-                    . ' for PHP ' . $phpVersion->toString()
-                    . ' with ' . $benchmarkConfiguration->toString()
-            );
-
-            $this->initBenchmark($phpVersion, $benchmarkConfiguration);
-
-            $body = $this->callUrl(BenchmarkUrlService::getPhpinfoUrl());
-            $this->outputSuccess('Http code is 200.');
-
-            if (is_string($body) === false || strlen($body) === 0) {
-                throw new \Exception('phpinfo() should not output an empty string.');
-            }
-            $this->outputSuccess('Response body is not empty.');
+    protected function afterHttpCodeValidated(
+        PhpVersion $phpVersion,
+        BenchmarkConfiguration $benchmarkConfiguration,
+        ?string $body
+    ): self {
+        if (is_string($body) === false || strlen($body) === false) {
+            throw new \Exception('phpinfo() should not output an empty string.');
         }
 
         return $this;
-    }
-
-    protected function initBenchmark(PhpVersion $phpVersion, BenchmarkConfiguration $benchmarkConfiguration): self
-    {
-        return $this->runCommand(
-            BenchmarkInitCommand::getDefaultName(),
-            [
-                'phpVersion' => $phpVersion->toString(),
-                '--no-url-output' => true,
-                '--opcache-enabled' => $benchmarkConfiguration->isOpcacheEnabled(),
-                '--preload-enabled' => $benchmarkConfiguration->isPreloadEnabled()
-            ]
-        );
     }
 }
