+++ src/Command/Validate/Benchmark/ValidateBenchmarkResponseCommand.php
@@ -9,20 +9,14 @@ use App\{
     Benchmark\BenchmarkType,
     Benchmark\BenchmarkUrlService,
     BenchmarkConfiguration\BenchmarkConfiguration,
-    BenchmarkConfiguration\BenchmarkConfigurationService,
-    Command\AbstractCommand,
     Command\Behavior\CallUrlTrait,
-    Command\Behavior\ValidateCircleCiOption,
-    Command\Benchmark\BenchmarkInitCommand,
-    Command\Validate\Configuration\ValidateConfigurationCommand,
     PhpVersion\PhpVersion,
     Utils\Path
 };
 
-final class ValidateBenchmarkResponseCommand extends AbstractCommand
+final class ValidateBenchmarkResponseCommand extends AbstractValidateBenchmarkUrlCommand
 {
     use CallUrlTrait;
-    use ValidateCircleCiOption;
 
     /** @var string */
     protected static $defaultName = 'validate:benchmark:response';
@@ -31,52 +25,19 @@ final class ValidateBenchmarkResponseCommand extends AbstractCommand
     {
         parent::configure();
 
-        $this
-            ->setDescription('Validate benchmark response')
-            ->addOption('no-validate-configuration')
-            ->addValidateCircleCiOption($this->getDefinition());
+        $this->setDescription('Validate benchmark response');
     }
 
-    protected function doExecute(): int
+    protected function getUrl(): string
     {
-        if ($this->getInput()->getOption('no-validate-configuration') === false) {
-            $this->runCommand(
-                ValidateConfigurationCommand::getDefaultName(),
-                $this->appendValidateCircleCiOption($this->getInput())
-            );
-        }
-
-        foreach (Benchmark::getCompatiblesPhpVersions() as $phpVersion) {
-            $this->validateForPhpVersion($phpVersion);
-        }
-
-        return 0;
+        return BenchmarkUrlService::getUrl(true);
     }
 
-    protected function validateForPhpVersion(PhpVersion $phpVersion): self
-    {
-        foreach (BenchmarkConfigurationService::getAvailable($phpVersion) as $benchmarkConfiguration) {
-            $this->outputTitle(
-                'Validation of '
-                    . BenchmarkUrlService::getUrl(true)
-                    . ' for PHP ' . $phpVersion->toString()
-                    . ' with ' . $benchmarkConfiguration->toString()
-            );
-
-            $this->initBenchmark($phpVersion, $benchmarkConfiguration);
-
-            $body = $this->callUrl(BenchmarkUrlService::getUrl(true));
-
-            $this
-                ->outputSuccess('Http code is 200.')
-                ->validateBody($body, $phpVersion);
-        }
-
-        return $this;
-    }
-
-    protected function validateBody(string $body, PhpVersion $phpVersion): self
-    {
+    protected function afterHttpCodeValidated(
+        PhpVersion $phpVersion,
+        BenchmarkConfiguration $benchmarkConfiguration,
+        ?string $body
+    ): self {
         $validated = false;
         $responseBodyPath = Path::getResponseBodyPath($phpVersion);
 
@@ -96,18 +57,5 @@ final class ValidateBenchmarkResponseCommand extends AbstractCommand
         }
 
         return $this;
-    }
-
-    protected function initBenchmark(PhpVersion $phpVersion, BenchmarkConfiguration $benchmarkConfiguration): self
-    {
-        return $this->runCommand(
-            BenchmarkInitCommand::getDefaultName(),
-            [
-                'phpVersion' => $phpVersion->toString(),
-                '--no-url-output' => true,
-                '--opcache-enabled' => $benchmarkConfiguration->isOpcacheEnabled(),
-                '--preload-enabled' => $benchmarkConfiguration->isPreloadEnabled()
-            ]
-        );
     }
 }
